+++ log-writer/dfcti_log_writer.py
@@ -140,4 +140,4 @@ class Write_Logs:
 
 
 MachineID.Generate_Machine_ID()
-Write_Logs.Write_Process(60, 1)
+Write_Logs.Write_Process(1000, 1)