+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Latest updates 
 
+* Add alert class-module with support for mail alerts -> `dfcti_log_reader.py`
 * Add archive with the log writer script
 * Improves the pipeline for uploading artifact to the current release tag
 * Added correct versioning